Add files via upload
This was a project based on blockchain using a remix-Ethereum application for running the code for the land registry purpose. In this code inputted manually the land values and id and then after storing the values performed functions such as the transfer of transaction name or the owner name new name changes were done with the help of the above code.